feat: connect patient dashboard fragments to live API, fix layout/binding errors

+++ mobile/app/src/main/java/com/example/appointmed/features/patient/adapters/DoctorAdapter.kt
@@ -11,7 +11,7 @@ class DoctorAdapter(
     private val onSelect: (Doctor) -> Unit
 ) : RecyclerView.Adapter<DoctorAdapter.DoctorViewHolder>() {
 
-    private var selectedId: Int? = null
+    private var selectedId: Long? = null
 
     inner class DoctorViewHolder(val binding: ItemDoctorCardBinding) :
         RecyclerView.ViewHolder(binding.root)
