+++ mobile/app/src/main/java/com/example/appointmed/features/patient/models/Doctor.kt
@@ -1,6 +1,7 @@
 package com.example.appointmed.features.patient.models
+
 data class Doctor(
-    val id: Int,
+    val id: Long,
     val name: String,
     val specialization: String
 )
